continue refactoring the conversation api
